Add Testing env, split CI into CI and Staging
As noted by @olliahonen in #9 and documented practically since the dawn
of time in [this blog
post](http://futurice.com/blog/five-environments-you-cannot-develop-without),
there are legitimate reasons to have more environments. I think it
should be taken as understood that this sort of document does not fit
100% of the cases 100% of the time, therefore if someone has good
reasons they can have fewer environments.